requirements.txt
beautifulsoup4==4.3.2
Pillow==4.0.0
python-resize-image==1.1.11
requests==2.12.3
TwitterAPI==2.4.2
